[IMP CHANGES] Added logic to traverse the Taxonomy Tree and Create the Prefix Map
1. Added Breadth First Traversal of Taxonomy Tree
2. Now, we can create Prefix Map for a Taxonomy Tree
3. Added printing functions for the Prefix Map as well
4. Modified the DELIMITER String on basis of which we are identifying
keys for the Prefix Map.

+++ src/TaxonomyPrefixMapValue.java
@@ -3,7 +3,7 @@
  */
 public class TaxonomyPrefixMapValue
 {
-	private int mNodeID;			// Node Identifier
+	private long mNodeID;			// Node Identifier
 	private boolean mIsLast;		// Should we look for other strings as prefixes in the map?
 	
 	/**
@@ -12,15 +12,15 @@ public class TaxonomyPrefixMapValue
 	 * @param nodeId	// Node Identifier
 	 * @param isLast	// Should we look for other strings as prefixes in the map?
 	 */
-	public TaxonomyPrefixMapValue(int nodeId, boolean isLast)
+	public TaxonomyPrefixMapValue(long nodeId, boolean isLast)
 	{
 		this.mNodeID = nodeId;
 		this.mIsLast = isLast;
 	}
 	
 	// Getter & Setter for NodeID
-	public int getNodeId() { return mNodeID;	}
-	public void setNodeId(int nodeId) {	this.mNodeID = nodeId; }
+	public long getNodeId() { return mNodeID; }
+	public void setNodeId(long nodeId) { this.mNodeID = nodeId; }
 	
 	// Getter & Setter for isLast
 	public boolean isLast() { return mIsLast; }
